documentation

+++ ida_lib/tests/conftest.py
@@ -11,11 +11,13 @@ from ida_lib.core.pipeline_local_ops import GaussianNoisePipeline
 size = (100, 100, 3)
 samples = 10
 
-def image_np(dtype = np.uint8):
+
+def image_np(dtype=np.uint8):
     return np.random.randint(low=0, high=256, size=(size[0], size[1], 3), dtype=dtype)
 
-def segmap(img, dtype = np.uint8):
-    segmap = np.zeros((img.shape[0], img.shape[1], 1), dtype = dtype)
+
+def segmap(img, dtype=np.uint8):
+    segmap = np.zeros((img.shape[0], img.shape[1], 1), dtype=dtype)
     segmap[28:171, 35:485, 0] = 1
     segmap[10:25, 30:245, 0] = 2
     segmap[10:25, 70:385, 0] = 3
@@ -28,15 +30,18 @@ def points():
     number_of_points = 15
     return np.random.randint(1, size[0], number_of_points * 2).reshape(number_of_points, 2)
 
-def mask(img, dtype = np.uint8):
+
+def mask(img, dtype=np.uint8):
     mask_example1 = np.zeros((img.shape[0], img.shape[1], 1), dtype=dtype)
     mask_example1[0:50, 0:50] = 1
     return mask_example1
 
-def heatmap(img, dtype = np.uint8):
+
+def heatmap(img, dtype=np.uint8):
     heatmap_complete = np.zeros((img.shape[0], img.shape[1], 1))
     heatmap_complete[0:img.shape[0] // 4, 0:img.shape[1] // 4, 0] = 1
     return heatmap_complete
+
 
 @pytest.fixture
 def numpy_image():
@@ -57,14 +62,15 @@ def numpy_image_batch():
 
 @pytest.fixture
 def torch_image():
-    return { 'image': kornia.image_to_tensor(image_np())}
+    return {'image': kornia.image_to_tensor(image_np())}
 
 
 @pytest.fixture
 def numpy_float_all_elements_item():
     img = np.random.randint(low=0, high=256, size=(size[0], size[1], 3)).astype(np.float)
-    return {'image': img, 'keypoints': points(), 'segmap': segmap(img, dtype=np.float), 'mask': mask(img, dtype = np.float),
-            'heatmap': heatmap(img, dtype = np.float)}
+    return {'image': img, 'keypoints': points(), 'segmap': segmap(img, dtype=np.float),
+            'mask': mask(img, dtype=np.float),
+            'heatmap': heatmap(img, dtype=np.float)}
 
 
 @pytest.fixture
@@ -77,7 +83,8 @@ def numpy_image_and_points_batch():
 @pytest.fixture
 def numpy_float_all_elements_batch():
     img = np.random.randint(low=0, high=256, size=(100, 100, 3)).astype(np.float)
-    data = {'image': img, 'keypoints': points(), 'segmap': segmap(img, dtype=np.float), 'mask': mask(img, dtype = np.float),
+    data = {'image': img, 'keypoints': points(), 'segmap': segmap(img, dtype=np.float),
+            'mask': mask(img, dtype=np.float),
             'heatmap': heatmap(img, dtype=np.float)}
     batch = [data.copy() for _ in range(samples)]
     return batch
@@ -86,29 +93,29 @@ def numpy_float_all_elements_batch():
 @pytest.fixture
 def torch_float_all_elements_item():
     img = np.random.randint(low=0, high=256, size=(100, 100, 3)).astype(np.float)
-    return {'image':  kornia.image_to_tensor(img),
+    return {'image': kornia.image_to_tensor(img),
             'keypoints': points(),
-            'segmap': kornia.image_to_tensor(segmap(img, dtype = np.float)),
-            'mask':  kornia.image_to_tensor(mask(img, dtype=np.float)),
+            'segmap': kornia.image_to_tensor(segmap(img, dtype=np.float)),
+            'mask': kornia.image_to_tensor(mask(img, dtype=np.float)),
             'heatmap': kornia.image_to_tensor(heatmap(img, dtype=np.float))}
 
 
 @pytest.fixture
 def numpy_item_without_image():
     img = np.random.randint(low=0, high=256, size=(size[0], size[1], 3)).astype(np.float)
-    return {'segmap': segmap(img, dtype=np.float), 'mask': mask(img, dtype = np.float)}
+    return {'segmap': segmap(img, dtype=np.float), 'mask': mask(img, dtype=np.float)}
 
 
 @pytest.fixture
 def numpy_item_2_mask():
     img = np.random.randint(low=0, high=256, size=(size[0], size[1], 3)).astype(np.float)
-    return {'image': img, 'mask': mask(img, dtype = np.float), 'mask2': mask(img, dtype = np.float)}
+    return {'image': img, 'mask': mask(img, dtype=np.float), 'mask2': mask(img, dtype=np.float)}
 
 
 @pytest.fixture
 def numpy_batch_2_mask():
     img = np.random.randint(low=0, high=256, size=(size[0], size[1], 3)).astype(np.float)
-    data = {'image': img, 'mask': mask(img, dtype = np.float), 'mask2': mask(img, dtype = np.float)}
+    data = {'image': img, 'mask': mask(img, dtype=np.float), 'mask2': mask(img, dtype=np.float)}
     batch = [data.copy() for _ in range(samples)]
     return batch
 
@@ -116,7 +123,7 @@ def numpy_batch_2_mask():
 @pytest.fixture
 def numpy_batch_without_image():
     img = np.random.randint(low=0, high=256, size=(size[0], size[1], 3)).astype(np.float)
-    data = {'segmap': segmap(img, dtype = np.float), 'mask': mask(img, dtype = np.float)}
+    data = {'segmap': segmap(img, dtype=np.float), 'mask': mask(img, dtype=np.float)}
     batch = [data.copy() for _ in range(samples)]
     return batch
